refactor(mini-window): extract state machine into hook
Phase 5 of frontend architecture refactor. Extracts all Tauri event
listeners and state (audioLevel, isRecording, recordingTime, status,
translateMode) into useMiniWindowState hook (171 lines). The
mini-window.tsx component drops from 295 to 169 lines of pure UI.

Final audit: no file exceeds 305 lines (excluding shadcn ui/). The
three original monoliths (setting-tabs 1370, notes-editor 910,
Dashboard 584) have been fully decomposed into feature folders with
isolated hooks.

All five refactor phases are now complete. Progress tracked in
docs/REFACTOR_FRONTEND_PROGRESS.md.
